requirements.txt
absl-py==2.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
beautifulsoup4==4.13.3
causal_conv1d==1.5.0.post8
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
docker-pycreds==0.4.0
einops==0.8.1
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.13.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.29.3
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.1
lightning-utilities==0.14.2
llvmlite==0.44.0
mamba_ssm==2.2.4
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
ml_collections==1.0.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.5.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.24.0
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchinfo==1.8.0
torchmetrics==1.7.0
torchvision==0.20.1+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wandb==0.19.8
wcwidth==0.2.13
widgetsnbextension==4.0.13
yarl==1.18.3
